+++ src/main/java/insane96mcp/enhancedai/mixin/MagmaCubeMixin.java
@@ -1,0 +1,25 @@
+package insane96mcp.enhancedai.mixin;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import insane96mcp.enhancedai.modules.slime.MagmaCubeSurfSpeed;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.monster.MagmaCube;
+import net.minecraft.world.entity.monster.Slime;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(MagmaCube.class)
+public class MagmaCubeMixin extends Slime {
+
+    public MagmaCubeMixin(EntityType<? extends Slime> pEntityType, Level pLevel) {
+        super(pEntityType, pLevel);
+    }
+
+    @WrapOperation(method = "jumpInLiquidInternal", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/monster/MagmaCube;setDeltaMovement(DDD)V"))
+    public void onJumpInLiquidInternal(MagmaCube instance, double x, double y, double z, Operation<Void> original) {
+        double multiplier = MagmaCubeSurfSpeed.getSpeedMultiplier();
+        original.call(instance, x * multiplier, y, z * multiplier);
+    }
+}